backend/requirements.txt
pymongo
strawberry-graphql
strawberry-graphql[cli]
strawberry-graphql[fastapi]
strawberry-graphql[artwork]
strawberry-graphql[strawberry]
fastapi
uvicorn[standard]
pydantic
redis
prometheus-client
psutil
requests
datetime
pytz